docs: record scoped stryker decision
