+++ src/Arius.Core.Tests/Shared/ChunkStorage/ChunkStorageServiceUploadTests.cs
@@ -313,165 +313,4 @@ public class ChunkStorageServiceUploadTests
         metadata.Metadata[BlobMetadataKeys.AriusType].ShouldBe(BlobMetadataKeys.TypeThin);
     }
 
-    private sealed class BlobAlreadyExistsOnSetMetadataOnceBlobContainerService : IBlobContainerService
-    {
-        private readonly FakeInMemoryBlobContainerService _inner = new();
-        private int _remainingFailures = 1;
-
-        public ICollection<string> DeletedBlobNames => _inner.DeletedBlobNames;
-
-        public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) =>
-            _inner.CreateContainerIfNotExistsAsync(cancellationToken);
-
-        public Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default) =>
-            _inner.UploadAsync(blobName, content, metadata, tier, contentType, overwrite, cancellationToken);
-
-        public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default) =>
-            _inner.OpenWriteAsync(blobName, contentType, cancellationToken);
-
-        public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.DownloadAsync(blobName, cancellationToken);
-
-        public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.GetMetadataAsync(blobName, cancellationToken);
-
-        public IAsyncEnumerable<string> ListAsync(string prefix, CancellationToken cancellationToken = default) =>
-            _inner.ListAsync(prefix, cancellationToken);
-
-        public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default)
-        {
-            if (Interlocked.Exchange(ref _remainingFailures, 0) == 1)
-                throw new BlobAlreadyExistsException(blobName);
-
-            return _inner.SetMetadataAsync(blobName, metadata, cancellationToken);
-        }
-
-        public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) =>
-            _inner.SetTierAsync(blobName, tier, cancellationToken);
-
-        public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) =>
-            _inner.CopyAsync(sourceBlobName, destinationBlobName, destinationTier, rehydratePriority, cancellationToken);
-
-        public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.DeleteAsync(blobName, cancellationToken);
-    }
-
-    private sealed class ThinMetadataRecordingBlobContainerService : IBlobContainerService
-    {
-        private readonly FakeInMemoryBlobContainerService _inner = new();
-
-        public List<string> UploadedBlobNames { get; } = [];
-        public Dictionary<string, IReadOnlyDictionary<string, string>> UploadMetadataByBlob { get; } = new(StringComparer.Ordinal);
-        public Dictionary<string, IReadOnlyDictionary<string, string>> SetMetadataByBlob { get; } = new(StringComparer.Ordinal);
-
-        public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) =>
-            _inner.CreateContainerIfNotExistsAsync(cancellationToken);
-
-        public Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default)
-        {
-            UploadedBlobNames.Add(blobName);
-            UploadMetadataByBlob[blobName] = new Dictionary<string, string>(metadata);
-            return _inner.UploadAsync(blobName, content, metadata, tier, contentType, overwrite, cancellationToken);
-        }
-
-        public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default) =>
-            _inner.OpenWriteAsync(blobName, contentType, cancellationToken);
-
-        public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.DownloadAsync(blobName, cancellationToken);
-
-        public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.GetMetadataAsync(blobName, cancellationToken);
-
-        public IAsyncEnumerable<string> ListAsync(string prefix, CancellationToken cancellationToken = default) =>
-            _inner.ListAsync(prefix, cancellationToken);
-
-        public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default)
-        {
-            SetMetadataByBlob[blobName] = new Dictionary<string, string>(metadata);
-            return _inner.SetMetadataAsync(blobName, metadata, cancellationToken);
-        }
-
-        public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) =>
-            _inner.SetTierAsync(blobName, tier, cancellationToken);
-
-        public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) =>
-            _inner.CopyAsync(sourceBlobName, destinationBlobName, destinationTier, rehydratePriority, cancellationToken);
-
-        public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.DeleteAsync(blobName, cancellationToken);
-    }
-
-    private sealed class ContentTypeCapturingBlobContainerService : IBlobContainerService
-    {
-        private readonly FakeInMemoryBlobContainerService _inner = new();
-        private readonly List<string?> _openWriteContentTypes = [];
-
-        public string? LastOpenWriteContentType { get; private set; }
-        public IReadOnlyList<string?> OpenWriteContentTypes => _openWriteContentTypes;
-
-        public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) =>
-            _inner.CreateContainerIfNotExistsAsync(cancellationToken);
-
-        public Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default) =>
-            _inner.UploadAsync(blobName, content, metadata, tier, contentType, overwrite, cancellationToken);
-
-        public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default)
-        {
-            LastOpenWriteContentType = contentType;
-            _openWriteContentTypes.Add(contentType);
-            return _inner.OpenWriteAsync(blobName, contentType, cancellationToken);
-        }
-
-        public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.DownloadAsync(blobName, cancellationToken);
-
-        public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.GetMetadataAsync(blobName, cancellationToken);
-
-        public IAsyncEnumerable<string> ListAsync(string prefix, CancellationToken cancellationToken = default) =>
-            _inner.ListAsync(prefix, cancellationToken);
-
-        public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default) =>
-            _inner.SetMetadataAsync(blobName, metadata, cancellationToken);
-
-        public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) =>
-            _inner.SetTierAsync(blobName, tier, cancellationToken);
-
-        public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) =>
-            _inner.CopyAsync(sourceBlobName, destinationBlobName, destinationTier, rehydratePriority, cancellationToken);
-
-        public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) =>
-            _inner.DeleteAsync(blobName, cancellationToken);
-    }
-
-    private sealed class EncryptedPassthroughService : IEncryptionService
-    {
-        public bool IsEncrypted => true;
-
-        public Stream WrapForEncryption(Stream inner) => inner;
-
-        public Stream WrapForDecryption(Stream inner) => inner;
-
-        public ContentHash ComputeHash(byte[] data) => throw new NotSupportedException();
-
-        public Task<ContentHash> ComputeHashAsync(Stream data, CancellationToken cancellationToken = default) => throw new NotSupportedException();
-
-        public Task<ContentHash> ComputeHashAsync(string filePath, IProgress<long>? progress = null, CancellationToken cancellationToken = default) => throw new NotSupportedException();
-    }
-
-    private sealed class ChunkedReadMemoryStream(byte[] buffer, int maxChunkSize) : MemoryStream(buffer, writable: false)
-    {
-        public override int Read(byte[] buffer, int offset, int count) =>
-            base.Read(buffer, offset, Math.Min(count, maxChunkSize));
-
-        public override int Read(Span<byte> buffer) =>
-            base.Read(buffer[..Math.Min(buffer.Length, maxChunkSize)]);
-
-        public override Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken) =>
-            base.ReadAsync(buffer, offset, Math.Min(count, maxChunkSize), cancellationToken);
-
-        public override ValueTask<int> ReadAsync(Memory<byte> buffer, CancellationToken cancellationToken = default) =>
-            base.ReadAsync(buffer[..Math.Min(buffer.Length, maxChunkSize)], cancellationToken);
-    }
 }
